rename parse_args.py to args.py to make room for more args functions
